requirements.txt
anthropic==0.40.0
httpx==0.27.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1
python-dotenv==1.0.0
requests==2.32.3

# LLM Provider Libraries
openai==1.54.0  # For OpenAI or Ollama
google-generativeai==0.8.3  # For Gemini
